requirements.txt
Pillow
tk
pyinstaller